+++ packages/generator/src/action-function-import/import.ts
@@ -12,6 +12,7 @@ import {
   externalImportDeclarations,
   mergeImportDeclarations
 } from '../imports';
+import { isEntityNotDeserializable } from '../edmx-to-vdm/common';
 import { responseTransformerFunctionName } from './response-transformer-function';
 
 function actionFunctionImportDeclarations(
@@ -25,8 +26,7 @@ function actionFunctionImportDeclarations(
     coreImportDeclaration([
       ...corePropertyTypeImportNames(parameters),
       ...returnTypes.map(returnType =>
-        returnType.returnTypeCategory ===
-        VdmReturnTypeCategory.ENTITY_NOT_DESERIALIZABLE
+        isEntityNotDeserializable(returnType)
           ? 'throwErrorWhenReturnTypeIsUnionType'
           : responseTransformerFunctionName(returnType, oDataVersion)
       ),
@@ -71,8 +71,7 @@ function returnTypeImports(
         returnType =>
           returnType.returnTypeCategory !== VdmReturnTypeCategory.EDM_TYPE &&
           returnType.returnTypeCategory !== VdmReturnTypeCategory.VOID &&
-          returnType.returnTypeCategory !==
-            VdmReturnTypeCategory.ENTITY_NOT_DESERIALIZABLE
+          returnType.returnTypeCategory !== VdmReturnTypeCategory.NEVER
       )
       .map(returnType => returnTypeImport(returnType))
   );